requirements.txt
streamlit
google-generativeai
speech_recognition
gtts
pydub